requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
dataclasses-json==0.5.14
extra-streamlit-components==0.1.60
frozenlist==1.4.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.261
langsmith==0.0.20
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
pandas==2.0.3
Pillow==9.5.0
plotly==5.16.1
polygon-api-client==1.12.3
protobuf==4.24.0
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
PyJWT==2.8.0
Pympler==1.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
streamlit==1.26.0
streamlit-authenticator==0.2.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.66.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2
